move GitException to Exception dir, move GitProcess to Process dir
