--- packages/GitWrapper/src/GitException.php
@@ -1,12 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Symplify\GitWrapper;
-
-use RuntimeException;
-
-/**
- * Base class for exceptions thrown by the Git library.
- */
-final class GitException extends RuntimeException
-{
-}
