+++ packages/GitWrapper/tests/AbstractGitWrapperTestCase.php
@@ -5,7 +5,7 @@ namespace Symplify\GitWrapper\Tests;
 use Nette\Utils\Random;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
-use Symplify\GitWrapper\GitException;
+use Symplify\GitWrapper\Exception\GitException ;
 use Symplify\GitWrapper\GitWrapper;
 
 abstract class AbstractGitWrapperTestCase extends TestCase
